# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trading Dashboard API - real-time market data, crypto, macro indicators, and news
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Quote {
  symbol?: string;
  name?: string;
  price?: number;
  change?: number;
  changePct?: number;
  volume?: number;
  open?: number;
  high?: number;
  low?: number;
  prevClose?: number;
  week52High?: number;
  week52Low?: number;
  marketCap?: number;
  pe?: number;
  avgVolume?: number;
  bid?: number;
  ask?: number;
}

export interface QuotesResponse {
  quotes?: Quote[];
  source?: string;
  asOf?: string;
}

export interface ChartBar {
  time?: string;
  open?: number;
  high?: number;
  low?: number;
  close?: number;
  volume?: number;
}

export interface ChartResponse {
  bars?: ChartBar[];
  symbol?: string;
  source?: string;
  asOf?: string;
}

export interface FinancialData { [key: string]: unknown }

export interface KeyStatistics { [key: string]: unknown }

export type FinancialSummaryIncomeItem = { [key: string]: unknown };

export type FinancialSummaryBalanceItem = { [key: string]: unknown };

export type FinancialSummaryCashflowItem = { [key: string]: unknown };

export interface FinancialSummary {
  symbol?: string;
  financialData?: FinancialData;
  keyStatistics?: KeyStatistics;
  income?: FinancialSummaryIncomeItem[];
  balance?: FinancialSummaryBalanceItem[];
  cashflow?: FinancialSummaryCashflowItem[];
  source?: string;
  asOf?: string;
}

export interface Option {
  strike?: number;
  /** @nullable */
  expiration?: string | null;
  lastPrice?: number;
  bid?: number;
  ask?: number;
  change?: number;
  changePct?: number;
  volume?: number;
  openInterest?: number;
  impliedVol?: number;
  inTheMoney?: boolean;
}

export interface OptionsChain {
  symbol?: string;
  underlyingPrice?: number;
  expirationDates?: string[];
  calls?: Option[];
  puts?: Option[];
  source?: string;
  asOf?: string;
}

export interface NewsItem {
  id?: string;
  headline?: string;
  /** @nullable */
  summary?: string | null;
  publisher?: string;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  publishedAt?: string | null;
  symbols?: string[];
  /** @nullable */
  thumbnail?: string | null;
  source?: string;
}

export interface NewsResponse {
  news?: NewsItem[];
  count?: number;
  dataSource?: string;
  asOf?: string;
}

export interface CryptoTicker {
  symbol?: string;
  price?: number;
  change?: number;
  changePct?: number;
  volume?: number;
  quoteVolume?: number;
  high?: number;
  low?: number;
}

export interface CryptoTickersResponse {
  tickers?: CryptoTicker[];
  source?: string;
  asOf?: string;
}

export interface KlineBar {
  time?: string;
  open?: number;
  high?: number;
  low?: number;
  close?: number;
  volume?: number;
}

export interface KlinesResponse {
  symbol?: string;
  interval?: string;
  bars?: KlineBar[];
  count?: number;
  source?: string;
  asOf?: string;
}

export type FundingRateRateHistoryItem = {
  time?: string;
  fundingRate?: number;
};

export interface FundingRate {
  symbol?: string;
  markPrice?: number;
  indexPrice?: number;
  fundingRate?: number;
  /** @nullable */
  nextFundingTime?: string | null;
  openInterest?: number;
  rateHistory?: FundingRateRateHistoryItem[];
  source?: string;
  asOf?: string;
}

export interface YieldCurveMaturity {
  maturity?: string;
  /** @nullable */
  current?: number | null;
  /** @nullable */
  monthAgo?: number | null;
  /** @nullable */
  yearAgo?: number | null;
}

export type YieldCurveSpreadCurrent = {
  date?: string;
  value?: number;
};

export type YieldCurveSpreadHistoryItem = {
  date?: string;
  value?: number;
};

export interface YieldCurveSpread {
  label?: string;
  current?: YieldCurveSpreadCurrent;
  history?: YieldCurveSpreadHistoryItem[];
}

export type YieldCurveMaturitiesItem = { [key: string]: unknown };

export interface YieldCurve {
  snapshot?: YieldCurveMaturity[];
  maturities?: YieldCurveMaturitiesItem[];
  spreads?: YieldCurveSpread[];
  source?: string;
  asOf?: string;
}

export type MacroIndicatorHistoryItem = {
  date?: string;
  value?: number;
};

export interface MacroIndicator {
  id?: string;
  label?: string;
  unit?: string;
  /** @nullable */
  value?: number | null;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  prevValue?: number | null;
  /** @nullable */
  change?: number | null;
  /** @nullable */
  yoyChange?: number | null;
  history?: MacroIndicatorHistoryItem[];
}

export interface MacroDashboard {
  indicators?: MacroIndicator[];
  source?: string;
  asOf?: string;
}

export type GetQuotesParams = {
/**
 * Comma-separated list of stock symbols (e.g., AAPL,NVDA,MSFT)
 */
symbols: string;
};

export type GetChartParams = {
interval?: string;
range?: string;
};

export type GetOptionsChainParams = {
date?: string;
};

export type GetNewsParams = {
symbols?: string;
count?: number;
};

export type GetCryptoTickersParams = {
symbols?: string;
};

export type GetCryptoKlinesParams = {
interval?: string;
limit?: number;
};